requirements.txt
numpy
RPi.GPIO
pillow
